backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
google-auth
google-auth-oauthlib
google-api-python-client
beautifulsoup4
google-genai